+++ apps/towbar-web-app/src/components/auto-deploy-control.tsx
@@ -1,5 +1,5 @@
 "use client";
-import { FloppyDiskIcon, Rocket01Icon } from "@hugeicons/core-free-icons";
+import { FloppyDiskIcon } from "@hugeicons/core-free-icons";
 
 import { HugeiconsIcon } from "@hugeicons/react";
 
@@ -75,9 +75,7 @@ export function AutoDeployControlEditor({
       <form onSubmit={save}>
         <Widget>
           <Widget.Header>
-            <Widget.Title icon={<HugeiconsIcon icon={Rocket01Icon} />}>
-              Auto-deploy
-            </Widget.Title>
+            <Widget.Title help={false}>Auto-deploy</Widget.Title>
           </Widget.Header>
           <Widget.Content className="content-grid">
             <Switch